Teleporting between background doors is possible
Background doors need to be marked on the event layer with a unique
tile.
The tile must contain an event property, with the name consisting of
"door" followed by an id. an example in test.map is "doorA".
Only use a door tile twice, sa the system is only designed to cooperate
with two way doors right now.